Replace myKeys.json with environment variables
Injecting secrets via environment variables seems to be best practice.
This commit therefore replaces the use of myKeys.json with environment variables.

This also solves the issue of the build breaking when myKeys.json is absent.

To allow easy management of local secrets for development and test, this commit also introduces the custom_env package https://github.com/erisanolasheni/custom-env.
This package provides:

- Supports reading environment variables from .env files
- Supports separate .env files for each environment, e.g. dev and test, by having separate .env.<Environment name> files.