Output Mocha test report and publish in Github Actions
Configure test runner to use the JSON file reporter and output the test report to mocha-test-report.json.
NB: The json-file-reporter does not allow to specify a directory, only a filename, so the report is unfortunately placed in the root folder.

Publish the test report in Github Actions so it's easy to see which tests caused the test step to fail.