Run Hardhat mainnet fork embedded in test-mainnet.js
Before, it was necessary to run the forked mainnet in a separate process. By embedding the network in the test process, execution of the test becomes simpler to execute.

As part of this, Hardhat configuration parameters has been moved to the configuration files in /config.
This allows easy environment specific/local overrides of the defaults.