Scripts for building and testing package and dep's
The 'build' script in each package now also recursively builds its dependencies in topological order.
The build of the package itself must now be in the 'build-this-package' script.

The 'test' script in each package still only tests the package, but there is now a 'test-with-dependencies' script, that also recursively runs tests in its dependencies in topological order.