Update offer filling logic, see description
Formerly if `offerWants=0`, a taker with `fillWants=true` would still
receive more than they asked for due to the `|| offerWants == 0` part of
the modified `if`.

This does not respect the intended spec of `fillWants` which is to give
at most `takerWants` to the taker.

We remove the disjunct from the conditional and update the case where
`fillWants=false` to avoid division by 0.