mangrove.js: Ensure offer list events are not missed in initialization
There was a risk that the offer list cache (=semibook) could become inconsistent with the on-chain offer list due to events being missed during initialization. The reason was, that event listeners were only registered after the cache had been initialized.

Now, the cache is initialized in the first callback and any subsequent events are queued until initialization completes.