mangrove.js: Fire all Market events after initilization
Market was not firing events for the first block for which is received a Mangrove offer list event. This was due to the offer list being fetched on the first received event, based on that event's block number. As event are only fired when the offer list cache is updated, no events were fired for that block, where the cache was up-to-date.

This fix fetches the order book from the block before the one that generated the first event which allows all events from later blocks to be fired.

Furthermore, a race condition in the initialization logic allowed events to sometimes be processed out of order causing the semibook logic to silently fail.

Finally, an bug in the integration test for market has been fixed.